+++ plugin-agave/src/bin/config-check.rs
@@ -13,6 +13,13 @@ struct Args {
 }
 
 fn main() -> anyhow::Result<()> {
+    anyhow::ensure!(
+        rustls::crypto::aws_lc_rs::default_provider()
+            .install_default()
+            .is_ok(),
+        "failed to call CryptoProvider::install_default()"
+    );
+
     let args = Args::parse();
     let _config = Config::load_from_file(args.config)?;
     println!("Config is OK!");